requirements.txt
appdirs==1.4.4
Brotli==1.0.9
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
Flask==2.2.2
fonttools==4.38.0
fs==2.4.16
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.1
MarkupSafe==2.1.1
python-dotenv==0.21.0
requests==2.28.1
six==1.16.0
unicodedata2==15.0.0
urllib3==1.26.12
Werkzeug==2.2.2
zopfli==0.2.1
